+++ src/llama_stack_client/_client.py
@@ -41,6 +41,7 @@ from .resources import (
     toolgroups,
     vector_dbs,
     completions,
+    vector_stores,
     scoring_functions,
     synthetic_data_generation,
 )
@@ -57,7 +58,6 @@ from .resources.agents import agents
 from .resources.responses import responses
 from .resources.tool_runtime import tool_runtime
 from .resources.post_training import post_training
-from .resources.vector_stores import vector_stores
 
 __all__ = [
     "Timeout",
