--- src/llama_stack_client/resources/vector_stores/__init__.py
@@ -1,33 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-from .files import (
-    FilesResource,
-    AsyncFilesResource,
-    FilesResourceWithRawResponse,
-    AsyncFilesResourceWithRawResponse,
-    FilesResourceWithStreamingResponse,
-    AsyncFilesResourceWithStreamingResponse,
-)
-from .vector_stores import (
-    VectorStoresResource,
-    AsyncVectorStoresResource,
-    VectorStoresResourceWithRawResponse,
-    AsyncVectorStoresResourceWithRawResponse,
-    VectorStoresResourceWithStreamingResponse,
-    AsyncVectorStoresResourceWithStreamingResponse,
-)
-
-__all__ = [
-    "FilesResource",
-    "AsyncFilesResource",
-    "FilesResourceWithRawResponse",
-    "AsyncFilesResourceWithRawResponse",
-    "FilesResourceWithStreamingResponse",
-    "AsyncFilesResourceWithStreamingResponse",
-    "VectorStoresResource",
-    "AsyncVectorStoresResource",
-    "VectorStoresResourceWithRawResponse",
-    "AsyncVectorStoresResourceWithRawResponse",
-    "VectorStoresResourceWithStreamingResponse",
-    "AsyncVectorStoresResourceWithStreamingResponse",
-]
