--- src/llama_stack_client/resources/vector_stores/files.py
@@ -1,201 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
-
-from __future__ import annotations
-
-from typing import Dict, Union, Iterable
-
-import httpx
-
-from ..._types import NOT_GIVEN, Body, Query, Headers, NotGiven
-from ..._utils import maybe_transform, async_maybe_transform
-from ..._compat import cached_property
-from ..._resource import SyncAPIResource, AsyncAPIResource
-from ..._response import (
-    to_raw_response_wrapper,
-    to_streamed_response_wrapper,
-    async_to_raw_response_wrapper,
-    async_to_streamed_response_wrapper,
-)
-from ..._base_client import make_request_options
-from ...types.vector_stores import file_create_params
-from ...types.vector_stores.vector_store_file import VectorStoreFile
-
-__all__ = ["FilesResource", "AsyncFilesResource"]
-
-
-class FilesResource(SyncAPIResource):
-    @cached_property
-    def with_raw_response(self) -> FilesResourceWithRawResponse:
-        """
-        This property can be used as a prefix for any HTTP method call to return
-        the raw response object instead of the parsed content.
-
-        For more information, see https://www.github.com/stainless-sdks/llama-stack-python#accessing-raw-response-data-eg-headers
-        """
-        return FilesResourceWithRawResponse(self)
-
-    @cached_property
-    def with_streaming_response(self) -> FilesResourceWithStreamingResponse:
-        """
-        An alternative to `.with_raw_response` that doesn't eagerly read the response body.
-
-        For more information, see https://www.github.com/stainless-sdks/llama-stack-python#with_streaming_response
-        """
-        return FilesResourceWithStreamingResponse(self)
-
-    def create(
-        self,
-        vector_store_id: str,
-        *,
-        file_id: str,
-        attributes: Dict[str, Union[bool, float, str, Iterable[object], object, None]] | NotGiven = NOT_GIVEN,
-        chunking_strategy: file_create_params.ChunkingStrategy | NotGiven = NOT_GIVEN,
-        # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
-        # The extra values given here take precedence over values defined on the client or passed to this method.
-        extra_headers: Headers | None = None,
-        extra_query: Query | None = None,
-        extra_body: Body | None = None,
-        timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
-    ) -> VectorStoreFile:
-        """
-        Attach a file to a vector store.
-
-        Args:
-          file_id: The ID of the file to attach to the vector store.
-
-          attributes: The key-value attributes stored with the file, which can be used for filtering.
-
-          chunking_strategy: The chunking strategy to use for the file.
-
-          extra_headers: Send extra headers
-
-          extra_query: Add additional query parameters to the request
-
-          extra_body: Add additional JSON properties to the request
-
-          timeout: Override the client-level default timeout for this request, in seconds
-        """
-        if not vector_store_id:
-            raise ValueError(f"Expected a non-empty value for `vector_store_id` but received {vector_store_id!r}")
-        return self._post(
-            f"/v1/openai/v1/vector_stores/{vector_store_id}/files",
-            body=maybe_transform(
-                {
-                    "file_id": file_id,
-                    "attributes": attributes,
-                    "chunking_strategy": chunking_strategy,
-                },
-                file_create_params.FileCreateParams,
-            ),
-            options=make_request_options(
-                extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
-            ),
-            cast_to=VectorStoreFile,
-        )
-
-
-class AsyncFilesResource(AsyncAPIResource):
-    @cached_property
-    def with_raw_response(self) -> AsyncFilesResourceWithRawResponse:
-        """
-        This property can be used as a prefix for any HTTP method call to return
-        the raw response object instead of the parsed content.
-
-        For more information, see https://www.github.com/stainless-sdks/llama-stack-python#accessing-raw-response-data-eg-headers
-        """
-        return AsyncFilesResourceWithRawResponse(self)
-
-    @cached_property
-    def with_streaming_response(self) -> AsyncFilesResourceWithStreamingResponse:
-        """
-        An alternative to `.with_raw_response` that doesn't eagerly read the response body.
-
-        For more information, see https://www.github.com/stainless-sdks/llama-stack-python#with_streaming_response
-        """
-        return AsyncFilesResourceWithStreamingResponse(self)
-
-    async def create(
-        self,
-        vector_store_id: str,
-        *,
-        file_id: str,
-        attributes: Dict[str, Union[bool, float, str, Iterable[object], object, None]] | NotGiven = NOT_GIVEN,
-        chunking_strategy: file_create_params.ChunkingStrategy | NotGiven = NOT_GIVEN,
-        # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
-        # The extra values given here take precedence over values defined on the client or passed to this method.
-        extra_headers: Headers | None = None,
-        extra_query: Query | None = None,
-        extra_body: Body | None = None,
-        timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
-    ) -> VectorStoreFile:
-        """
-        Attach a file to a vector store.
-
-        Args:
-          file_id: The ID of the file to attach to the vector store.
-
-          attributes: The key-value attributes stored with the file, which can be used for filtering.
-
-          chunking_strategy: The chunking strategy to use for the file.
-
-          extra_headers: Send extra headers
-
-          extra_query: Add additional query parameters to the request
-
-          extra_body: Add additional JSON properties to the request
-
-          timeout: Override the client-level default timeout for this request, in seconds
-        """
-        if not vector_store_id:
-            raise ValueError(f"Expected a non-empty value for `vector_store_id` but received {vector_store_id!r}")
-        return await self._post(
-            f"/v1/openai/v1/vector_stores/{vector_store_id}/files",
-            body=await async_maybe_transform(
-                {
-                    "file_id": file_id,
-                    "attributes": attributes,
-                    "chunking_strategy": chunking_strategy,
-                },
-                file_create_params.FileCreateParams,
-            ),
-            options=make_request_options(
-                extra_headers=extra_headers, extra_query=extra_query, extra_body=extra_body, timeout=timeout
-            ),
-            cast_to=VectorStoreFile,
-        )
-
-
-class FilesResourceWithRawResponse:
-    def __init__(self, files: FilesResource) -> None:
-        self._files = files
-
-        self.create = to_raw_response_wrapper(
-            files.create,
-        )
-
-
-class AsyncFilesResourceWithRawResponse:
-    def __init__(self, files: AsyncFilesResource) -> None:
-        self._files = files
-
-        self.create = async_to_raw_response_wrapper(
-            files.create,
-        )
-
-
-class FilesResourceWithStreamingResponse:
-    def __init__(self, files: FilesResource) -> None:
-        self._files = files
-
-        self.create = to_streamed_response_wrapper(
-            files.create,
-        )
-
-
-class AsyncFilesResourceWithStreamingResponse:
-    def __init__(self, files: AsyncFilesResource) -> None:
-        self._files = files
-
-        self.create = async_to_streamed_response_wrapper(
-            files.create,
-        )
