+++ tests/api_resources/test_embeddings.py
@@ -9,7 +9,7 @@ import pytest
 
 from tests.utils import assert_matches_type
 from llama_stack_client import LlamaStackClient, AsyncLlamaStackClient
-from llama_stack_client.types import EmbeddingsResponse
+from llama_stack_client.types import CreateEmbeddingsResponse
 
 base_url = os.environ.get("TEST_API_BASE_URL", "http://127.0.0.1:4010")
 
@@ -23,7 +23,7 @@ class TestEmbeddings:
             input="string",
             model="model",
         )
-        assert_matches_type(EmbeddingsResponse, embedding, path=["response"])
+        assert_matches_type(CreateEmbeddingsResponse, embedding, path=["response"])
 
     @parametrize
     def test_method_create_with_all_params(self, client: LlamaStackClient) -> None:
@@ -34,7 +34,7 @@ class TestEmbeddings:
             encoding_format="encoding_format",
             user="user",
         )
-        assert_matches_type(EmbeddingsResponse, embedding, path=["response"])
+        assert_matches_type(CreateEmbeddingsResponse, embedding, path=["response"])
 
     @parametrize
     def test_raw_response_create(self, client: LlamaStackClient) -> None:
@@ -46,7 +46,7 @@ class TestEmbeddings:
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         embedding = response.parse()
-        assert_matches_type(EmbeddingsResponse, embedding, path=["response"])
+        assert_matches_type(CreateEmbeddingsResponse, embedding, path=["response"])
 
     @parametrize
     def test_streaming_response_create(self, client: LlamaStackClient) -> None:
@@ -58,7 +58,7 @@ class TestEmbeddings:
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             embedding = response.parse()
-            assert_matches_type(EmbeddingsResponse, embedding, path=["response"])
+            assert_matches_type(CreateEmbeddingsResponse, embedding, path=["response"])
 
         assert cast(Any, response.is_closed) is True
 
@@ -72,7 +72,7 @@ class TestAsyncEmbeddings:
             input="string",
             model="model",
         )
-        assert_matches_type(EmbeddingsResponse, embedding, path=["response"])
+        assert_matches_type(CreateEmbeddingsResponse, embedding, path=["response"])
 
     @parametrize
     async def test_method_create_with_all_params(self, async_client: AsyncLlamaStackClient) -> None:
@@ -83,7 +83,7 @@ class TestAsyncEmbeddings:
             encoding_format="encoding_format",
             user="user",
         )
-        assert_matches_type(EmbeddingsResponse, embedding, path=["response"])
+        assert_matches_type(CreateEmbeddingsResponse, embedding, path=["response"])
 
     @parametrize
     async def test_raw_response_create(self, async_client: AsyncLlamaStackClient) -> None:
@@ -95,7 +95,7 @@ class TestAsyncEmbeddings:
         assert response.is_closed is True
         assert response.http_request.headers.get("X-Stainless-Lang") == "python"
         embedding = await response.parse()
-        assert_matches_type(EmbeddingsResponse, embedding, path=["response"])
+        assert_matches_type(CreateEmbeddingsResponse, embedding, path=["response"])
 
     @parametrize
     async def test_streaming_response_create(self, async_client: AsyncLlamaStackClient) -> None:
@@ -107,6 +107,6 @@ class TestAsyncEmbeddings:
             assert response.http_request.headers.get("X-Stainless-Lang") == "python"
 
             embedding = await response.parse()
-            assert_matches_type(EmbeddingsResponse, embedding, path=["response"])
+            assert_matches_type(CreateEmbeddingsResponse, embedding, path=["response"])
 
         assert cast(Any, response.is_closed) is True
