--- tests/api_resources/vector_stores/__init__.py
@@ -1,1 +1,0 @@
-# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
